Refactor after review and resolve conflicts
