+++ src/main/java/faang/school/achievement/config/RedisMQConfig.java
@@ -1,5 +1,6 @@
 package faang.school.achievement.config;
 
+import faang.school.achievement.listener.CommentEventListener;
 import faang.school.achievement.listener.achievement.AchievementListener;
 import faang.school.achievement.listener.post.PostEventListener;
 import org.springframework.beans.factory.annotation.Value;
@@ -17,14 +18,16 @@ import org.springframework.data.redis.serializer.StringRedisSerializer;
 
 @Configuration
 public class RedisMQConfig {
-    @Value("${spring.data.redis.channel.achievement}")
-    private String achievementTopicName;
-    @Value("${spring.data.redis.channel.post}")
-    private String postTopicName;
     @Value("${spring.data.redis.host}")
     private String host;
     @Value("${spring.data.redis.port}")
     private int port;
+    @Value("${spring.data.redis.channel.achievement}")
+    private String achievementTopicName;
+    @Value("${spring.data.redis.channel.post}")
+    private String postTopicName;
+    @Value("${spring.data.redis.channel.comment}")
+    private String commentTopicName;
 
 
     @Bean
@@ -33,11 +36,6 @@ public class RedisMQConfig {
         RedisStandaloneConfiguration configuration =
                 new RedisStandaloneConfiguration(host, port);
         return new LettuceConnectionFactory(configuration);
-    }
-
-    @Bean
-    public ChannelTopic achievementTopic() {
-        return new ChannelTopic(achievementTopicName);
     }
 
     @Bean
@@ -50,30 +48,47 @@ public class RedisMQConfig {
     }
 
     @Bean
+    public ChannelTopic achievementTopic() {
+        return new ChannelTopic(achievementTopicName);
+    }
+
+    @Bean
     MessageListenerAdapter achievementMessageListener(AchievementListener achievementListener) {
         return new MessageListenerAdapter(achievementListener);
     }
 
     @Bean
-    public MessageListenerAdapter redisPostEventListener(PostEventListener postEventListener) {
-        return new MessageListenerAdapter(postEventListener);
-    }
-
-    @Bean
-    public ChannelTopic postEventTopic() {
+    public ChannelTopic postTopic() {
         return new ChannelTopic(postTopicName);
     }
 
     @Bean
+    public MessageListenerAdapter postMessageListener(PostEventListener postEventListener) {
+        return new MessageListenerAdapter(postEventListener);
+    }
+
+    @Bean
+    ChannelTopic commentTopic() {
+        return new ChannelTopic(commentTopicName);
+    }
+
+    @Bean
+    MessageListenerAdapter commentMessageListener(CommentEventListener commentEventListener) {
+        return new MessageListenerAdapter(commentEventListener);
+    }
+
+    @Bean
     RedisMessageListenerContainer redisContainer(MessageListenerAdapter achievementMessageListener,
-                                                 MessageListenerAdapter redisPostEventListener) {
+                                                 MessageListenerAdapter postMessageListener,
+                                                 MessageListenerAdapter commentMessageListener) {
         RedisMessageListenerContainer container =
                 new RedisMessageListenerContainer();
 
         container.setConnectionFactory(connectionFactory());
 
         container.addMessageListener(achievementMessageListener, achievementTopic());
-        container.addMessageListener(redisPostEventListener, postEventTopic());
+        container.addMessageListener(postMessageListener, postTopic());
+        container.addMessageListener(commentMessageListener, commentTopic());
 
         return container;
     }
