--- src/main/java/faang/school/achievement/config/context/RedisConfig.java
@@ -1,67 +1,0 @@
-package faang.school.achievement.config.context;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import faang.school.achievement.listener.CommentEventListener;
-import faang.school.achievement.model.CommentEvent;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
-import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.RedisMessageListenerContainer;
-import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
-import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.StringRedisSerializer;
-
-@Slf4j
-@Configuration
-public class RedisConfig {
-
-    @Value("${spring.data.redis.host}")
-    private String host;
-    @Value("${spring.data.redis.port}")
-    private int port;
-    @Value("${spring.data.redis.channel.comment}")
-    private String commentChannelName;
-
-    @Bean
-    public JedisConnectionFactory redisConnectionFactory() {
-        log.info("Redis host:{}, port:{}", host, port);
-        RedisStandaloneConfiguration config = new RedisStandaloneConfiguration(host, port);
-        return new JedisConnectionFactory(config);
-    }
-
-    @Bean
-    MessageListenerAdapter followerListener(CommentEventListener commentEventListener) {
-        return new MessageListenerAdapter(commentEventListener);
-    }
-
-    @Bean
-    public RedisMessageListenerContainer container(CommentEventListener commentEventListener) {
-        RedisMessageListenerContainer container = new RedisMessageListenerContainer();
-        container.setConnectionFactory(redisConnectionFactory());
-        container.addMessageListener(commentEventListener, new ChannelTopic(commentChannelName));
-        return container;
-    }
-
-    @Bean
-    public RedisTemplate<String, Object> redisTemplate(RedisConnectionFactory redisConnectionFactory,
-                                                       ObjectMapper mapper) {
-        RedisTemplate<String, Object> redisTemplate = new RedisTemplate<>();
-        redisTemplate.setConnectionFactory(redisConnectionFactory);
-        redisTemplate.setKeySerializer(new StringRedisSerializer());
-        redisTemplate.setValueSerializer(new GenericJackson2JsonRedisSerializer(mapper));
-        return redisTemplate;
-    }
-
-    @Bean
-    ChannelTopic commentTopic() {
-        return new ChannelTopic(commentChannelName);
-    }
-
-}
