+++ src/main/java/faang/school/achievement/eventHandler/ExpertAchievementHandler.java
@@ -1,7 +1,7 @@
-package faang.school.achievement.eventHandler;
+package faang.school.achievement.eventhandler;
 
-import faang.school.achievement.cache.AchievementCache;
-import faang.school.achievement.service.AchievementService;
+import faang.school.achievement.service.AchievementCache;
+import faang.school.achievement.service.achievement.AchievementService;
 import org.springframework.stereotype.Component;
 
 @Component
