--- src/main/java/faang/school/achievement/service/AchievementService.java
@@ -1,55 +1,0 @@
-package faang.school.achievement.service;
-
-import faang.school.achievement.model.Achievement;
-import faang.school.achievement.model.AchievementProgress;
-import faang.school.achievement.model.UserAchievement;
-import faang.school.achievement.repository.AchievementProgressRepository;
-import faang.school.achievement.repository.AchievementRepository;
-import faang.school.achievement.repository.UserAchievementRepository;
-import jakarta.persistence.EntityNotFoundException;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class AchievementService {
-
-    private final UserAchievementRepository userAchievementRepository;
-    private final AchievementProgressRepository achievementProgressRepository;
-    private final AchievementRepository achievementRepository;
-
-    @Transactional(readOnly = true)
-    public boolean hasAchievement(Long userId, Long achievementId) {
-        return userAchievementRepository.existsByUserIdAndAchievementId(userId, achievementId);
-    }
-
-    @Transactional
-    public void createProgressIfNecessary(Long userId, Long achievementId) {
-        achievementProgressRepository.createProgressIfNecessary(userId, achievementId);
-        log.info("user started to complete achievement: {}", achievementId);
-    }
-
-    @Transactional(readOnly = true)
-    public AchievementProgress getProgress(Long userId, Long achievementId) {
-        return achievementProgressRepository.findByUserIdAndAchievementId(userId, achievementId)
-                .orElseThrow(()->new EntityNotFoundException("no progress was found"));
-    }
-
-    @Transactional
-    public void updateProgress(AchievementProgress progress) {
-        achievementProgressRepository.save(progress);
-    }
-
-    @Transactional
-    public void giveAchievement(Long userId, Achievement achievement) {
-        UserAchievement newAchievement = UserAchievement.builder()
-                .userId(userId)
-                .achievement(achievement)
-                .build();
-        log.info("user {} obtained achievement {}", userId, achievement);
-        userAchievementRepository.save(newAchievement);
-    }
-}
